Rollup config updates to support dynamic require and mixed es modules, blog search layout, restructured data for articleCard.
