+++ src/ja-JP/news/blog/search-output.html
@@ -4,6 +4,6 @@ permalink: '/{{ locale }}/news/blog/search-output.json'
 override:tags: []
 ---
 
-[ {% for item in collections['ja-JP-blog-post'] %} { "title":"{{item.data.title}}", "author":"{{item.data.author}}",
-"date":"{{item.data.date}}", "url":"{{item.url}}", "content":"{{ item.templateContent | removeHtml | truncate }}" }{% if loop.last != true
-%},{% endif %} {% endfor %} ]
+[ {% for item in collections['ja-JP-blog-post'] %} { "image":"{{item.data.image}}", "title":"{{item.data.title}}",
+"author":"{{item.data.author}}", "date":"{{item.data.date}}", "url":"{{item.url}}", "content":"{{ item.templateContent | removeHtml |
+truncate }}" }{% if loop.last != true %},{% endif %} {% endfor %} ]
