+++ src/js/search-output.js
@@ -1,79 +1,75 @@
 import lunr from 'lunr';
+import articleCard from '../_11ty/shortcodes/ArticleCard.js';
 
 const SearchOutput = {
   init: () => {
     const urlPath = window.location.pathname.split('/');
     const urlLocale = urlPath[1];
     const searchInput = document.getElementById('search-str');
-    const searchSubmit = document.getElementById('search-submit');
     const searchresultsContainer = document.getElementById('search-results');
     const queryString = window.location.search;
     const urlParams = new URLSearchParams(queryString);
     const query = urlParams.get('q');
 
-    let postsIndex, searchIndex, searchResults, searchResultsHtml;
-
     async function initSearchIndex() {
-      const response = await fetch(`/${urlLocale}/news/blog/search.json`);
-      postsIndex = await response.json();
-      searchIndex = lunr(function () {
-        this.field('title');
-        this.field('author');
-        this.field('date');
-        this.ref('url');
-        this.field('content');
-        postsIndex.forEach(post => this.add(post));
-      });
+      const searchIndexData = await fetch(
+        `/${urlLocale}/news/blog/search-index.json`
+      );
+      const index = await searchIndexData.json();
+      const lunrIndex = lunr.Index.load(index);
+
+      const searchOutputData = await fetch(
+        `/${urlLocale}/news/blog/search-output.json`
+      );
+      const output = await searchOutputData.json();
+
       if (!urlParams.has('q')) return;
-      search(query);
+      search(lunrIndex, query, output);
     }
 
-    function search(searchQuery) {
+    function search(lunrIndex, searchQuery, searchOutput) {
       searchInput.value = searchQuery;
 
-      const searchResults = searchIndex.search(searchQuery);
+      let searchResults = lunrIndex.search(query);
 
       searchResults.forEach(result => {
-        const matchingPost = postsIndex.find(post => post.url === result.ref);
-        if (!matchingPost) return;
-
-        const { title, author, date, url, content } = matchingPost;
-        const searchData = {
-          title,
-          author,
-          date,
-          url,
-          content,
-        };
-        result.data = searchData;
+        result.image = searchOutput[result.ref].image;
+        result.title = searchOutput[result.ref].title;
+        result.author = searchOutput[result.ref].author;
+        result.date = searchOutput[result.ref].date;
+        result.url = searchOutput[result.ref].url;
+        result.content = searchOutput[result.ref].content;
       });
 
       renderResults(searchResults, searchQuery);
     }
 
-    function formatDate(date, locale = 'en-GB') {
-      return new Intl.DateTimeFormat(locale, {
-        year: 'numeric',
-        month: 'long',
-        day: 'numeric',
-      }).format(new Date(date));
-    }
-
     function renderResults(results = [], searchQuery) {
+      let searchResultsHtml;
+
       if (!searchresultsContainer) return;
 
       if (!results.length) {
-        searchResultsHtml = `<p class="p">No results for "<strong>${searchQuery}</strong>"</p>`;
+        searchResultsHtml = `<p class="h3 mb-8 xl:mb-10">No results for “${searchQuery}”</p>`;
       } else {
-        searchResultsHtml = `<p class="p">You searched for "<strong>${searchQuery}</strong>"</p>
-      <ul class="ul">${results
-        .map(({ data = {} }) => {
-          const { author, date, title, url } = data;
-          return `<li>
-                  <a href="${url}">${title}</a>
-                  <span class="block">${author}</span>
-                  <time datetime="${date}">${formatDate(date)}</time>
-                </li>`;
+        searchResultsHtml = `<p class="h3 mb-8 xl:mb-10">You searched for “${searchQuery}”</p>
+      <ul class="grid md:grid--cols-2 lg:grid--cols-3 xl:grid--cols-4 list-none m-0 p-0">${results
+        .map(result => {
+          const { author, content, date, image, title, url } = result;
+
+          const restucturedData = {
+            data: {
+              author,
+              date,
+              image,
+              title,
+              locale: urlLocale,
+            },
+            templateContent: content,
+            url,
+          };
+
+          return `<li>${articleCard(restucturedData)}</li>`;
         })
         .join('')}</ul>`;
       }
